Pull CurrencyType out of CurrencyItem class
